# TensorAlgebra cleanup and optimization