--- a/src/PhysicalModels/ViscousModels.jl
+++ b/src/PhysicalModels/ViscousModels.jl
@@ -63,29 +63,6 @@ function updateStateVariables!(model::GeneralizedMaxwell, Δt, u, un, stateVars)
 end
 
 
-# """
-#   _getKinematic(::Visco, ∇u, Uv⁻¹)
-
-# Compute the kinematics of a viscous model.
-
-# # Arguments
-# - `::Visco`: A viscous model
-# - `∇u`: The deformation gradient at the considered time step
-# - `Uv⁻¹`: The inverse of the viscous strain at the considered time step
-
-# # Returns
-# - `F`
-# - `C`
-# - `Ce`
-# """
-# function _getKinematic(::Visco, ∇u, Uv⁻¹)
-#   F = one(∇u) + ∇u
-#   C = F' * F
-#   Ce = Uv⁻¹ * C * Uv⁻¹
-#   return (F, C, Ce)
-# end
-
-
 """
   ViscousStrain(Ce::TensorValue, C::TensorValue)::TensorValue
   
@@ -348,15 +325,9 @@ end
 function Piola(obj::ViscousIncompressible, Δt::Float64,
                 Se_::Function, ∂Se∂Ce_::Function,
                 F::TensorValue, Fn::TensorValue, stateVars::VectorValue)
-                # ∇u_::TensorValue, ∇un_::TensorValue, stateVars::VectorValue)
   state_vars = get_array(stateVars)
-  Uvn = TensorValue{3,3}(Tuple(state_vars[1:9]))
-  # Uvn = SMatrix{3,3}(state_vars[1:9])
+  Uvn = TensorValue{3,3}(Tuple(state_vars[1:9]))  # TODO: Update tensor slicing until next Gridap version has been released
   λαn = state_vars[10]
-#   ∇u = get_array(∇u_)
-#   ∇un = get_array(∇un_)
-  # F = get_array(F_)
-  # Fn = get_array(Fn_)
   #------------------------------------------
   # Get kinematics
   #------------------------------------------
@@ -366,24 +337,16 @@ function Piola(obj::ViscousIncompressible, Δt::Float64,
   Cn = C_(Fn)
   Ceᵗʳ = Ce_(C, invUvn)
   Cen  = Ce_(Cn, invUvn)
-#   F, C, Ceᵗʳ = _getKinematic(obj, ∇u, invUvn)
-#   _, _, Cen  = _getKinematic(obj, ∇un, invUvn)
   #------------------------------------------
   # Return mapping algorithm
   #------------------------------------------
   Ce, _ = return_mapping_algorithm!(obj, Δt, Se_, ∂Se∂Ce_, F, Ceᵗʳ, Cen, λαn)
   #------------------------------------------
-  # Get invUv and Sα
+  # Get invUv and Pα
   #------------------------------------------
   _, _, invUv = ViscousStrain(Ce, C)
   Pα = ViscousPiola(Se_, Ce, invUv, F)
-  #------------------------------------------
-  # Tangent operator
-  #------------------------------------------
-  # check_type = Pα isa TensorValue{3,3,Float64}
-  # if !check_type throw("Pα is a $(typeof(Pα))") end
   Pα
-  # return TensorValue(Pα)
 end
 
 
@@ -405,15 +368,9 @@ Visco-Elastic model for incompressible case
 function Tangent(obj::ViscousIncompressible, Δt::Float64,
                  Se_::Function, ∂Se∂Ce_::Function,
                  F::TensorValue, Fn::TensorValue, stateVars::VectorValue)
-                #  ∇u_::TensorValue, ∇un_::TensorValue, stateVars::VectorValue)
   state_vars = get_array(stateVars)
-  Uvn = TensorValue{3,3}(Tuple(state_vars[1:9]))
-  # Uvn = SMatrix{3,3}(state_vars[1:9])
+  Uvn = TensorValue{3,3}(Tuple(state_vars[1:9]))  # TODO: Update tensor slicing until next Gridap version has been released
   λαn = state_vars[10]
-#   ∇u = get_array(∇u_)
-#   ∇un = get_array(∇un_)
-#   F = get_array(F_)
-#   Fn = get_array(Fn_)
   #------------------------------------------
   # Get kinematics
   #------------------------------------------
@@ -423,8 +380,6 @@ function Tangent(obj::ViscousIncompressible, Δt::Float64,
   Cn = C_(Fn)
   Ceᵗʳ = Ce_(C, invUvn)
   Cen  = Ce_(Cn, invUvn)
-#   F, C, Ceᵗʳ = _getKinematic(obj, ∇u, invUvn)
-#   _, _, Cen  = _getKinematic(obj, ∇un, invUvn)
   #------------------------------------------
   # Return mapping algorithm
   #------------------------------------------
@@ -461,13 +416,8 @@ function ReturnMapping(obj::ViscousIncompressible, Δt::Float64,
                        Se_::Function, ∂Se∂Ce_::Function,
                        F::TensorValue, Fn::TensorValue, stateVars::VectorValue)
   state_vars = get_array(stateVars)
-  Uvn = TensorValue{3,3}(Tuple(state_vars[1:9]))
-  # Uvn = SMatrix{3,3}(state_vars[1:9])
+  Uvn = TensorValue{3,3}(Tuple(state_vars[1:9]))  # TODO: Update tensor slicing until next Gridap version has been released
   λαn = state_vars[10]
-  # ∇u = get_array(∇u_)
-  # ∇un = get_array(∇un_)
-  # F = get_array(F_)
-  # Fn = get_array(Fn_)
   #------------------------------------------
   # Get kinematics
   #------------------------------------------
@@ -477,8 +427,6 @@ function ReturnMapping(obj::ViscousIncompressible, Δt::Float64,
   Cn = C_(Fn)
   Ceᵗʳ = Ce_(C, invUvn)
   Cen  = Ce_(Cn, invUvn)
-#   F, C, Ceᵗʳ = _getKinematic(obj, ∇u, invUvn)
-#   _, _, Cen  = _getKinematic(obj, ∇un, invUvn)
   #------------------------------------------
   # Return mapping algorithm
   #------------------------------------------
@@ -496,13 +444,8 @@ function ViscousDissipation(obj::ViscousIncompressible, Δt::Float64,
                        Se_::Function, ∂Se∂Ce_::Function,
                        F::TensorValue, Fn::TensorValue, stateVars::VectorValue)
   state_vars = get_array(stateVars)
-  Uvn = TensorValue{3,3}(Tuple(state_vars[1:9]))
-  # Uvn = SMatrix{3,3}(state_vars[1:9])
+  Uvn = TensorValue{3,3}(Tuple(state_vars[1:9]))  # TODO: Update tensor slicing until next Gridap version has been released
   λαn = state_vars[10]
-  # ∇u = get_array(∇u_)
-  # ∇un = get_array(∇un_)
-  # F = get_array(F_)
-  # Fn = get_array(Fn_)
   #------------------------------------------
   # Get kinematics
   #------------------------------------------
@@ -512,8 +455,6 @@ function ViscousDissipation(obj::ViscousIncompressible, Δt::Float64,
   Cn = C_(Fn)
   Ceᵗʳ = Ce_(C, invUvn)
   Cen  = Ce_(Cn, invUvn)
-#   F, C, Ceᵗʳ = _getKinematic(obj, ∇u, invUvn)
-#   _, _, Cen  = _getKinematic(obj, ∇un, invUvn)
   #------------------------------------------
   # Return mapping algorithm
   #------------------------------------------